Moved Overdrive classifier into its own file.
